docker_airflow_kafka_pyspark/requirements.txt
openmeteo-requests
requests-cache 
retry-requests

kafka-python-ng
pyspark